Progress monitor implementation

+++ src/main/java/me/lumpchen/xdiff/ProgressListener.java
@@ -2,6 +2,6 @@ package me.lumpchen.xdiff;
 
 public interface ProgressListener {
 
-	public void progress(int progress);
+	public void progress(float progress);
 	
 }
